[Change]　レイヤーの調整

+++ ProjectData/GameSources/ObjectBase.cpp
@@ -47,11 +47,12 @@ namespace basecross {
 		auto &app = App::GetApp();
 		auto texDir = app->GetDataDirWString();
 
+		//アルファ値を有効にする
 		drawComp->SetTextureResource(m_textureName);
 		this->SetAlphaActive(true);
 
 		float Layer = m_position.z * (-1.0f);
-		SetDrawLayer(Layer);
+		SetDrawLayer((int)Layer);
 	}
 
 }
